pyproject.toml
 [tool.poetry]
name = "pubmed-pharma-papers"
version = "0.1.0"
description = "A tool to fetch research papers from PubMed with pharmaceutical/biotech company affiliations"
authors = ["vvsspraveen <jampanasaipraveen@gmail.com>"]
readme = "README.md"
packages = [{include = "pubmed_pharma_papers"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pandas = "^2.0.3"
biopython = "^1.81"
typing-extensions = "^4.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
mypy = "^1.3.0"
isort = "^5.12.0"

[tool.poetry.scripts]
get-papers-list = "pubmed_pharma_papers.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
